pyproject.toml
[tool.poetry]
name = "civitas"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aerich = "^0.7.2"
fastapi = "^0.111.0"
fastapi-cache2 = "^0.2.1"
fastapi-pagination = "^0.12.24"
google-cloud-bigquery = "^3.23.1"
googlemaps = "^4.10.0"
httpx = "^0.27.0"
infisical = "<1.6"
loguru = "^0.7.2"
orjson = "^3.10.3"
pendulum = "^3.0.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pytz = "^2024.1"
redis = "^5.0.4"
sentry-sdk = { extras = ["fastapi"], version = "^2.3.1" }
slowapi = "^0.1.9"
tortoise-orm = { version = "^0.20.0", extras = ["asyncpg"] }
uvicorn = { extras = ["standard"], version = "^0.30.0" }


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
pre-commit = "^3.7.1"
taskipy = "^1.12.2"
isort = "^5.13.2"

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "black . && isort . && flake8 ."
serve = "uvicorn app.main:app --reload --port 8080"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
